style(FacebookAuthenticationButton): only include a single file
